Move align_node_attrs to utils.py
